create middleware role admin

+++ routes/web.php
@@ -23,7 +23,7 @@ Route::get('/', function () {
 })->middleware('auth');
 
 Route::get('/login', [AuthController::class, 'login'])->name('login')->middleware('guest');
-Route::post('/login', [AuthController::class, 'authentication']);
+Route::post('/login', [AuthController::class, 'authentication'])->middleware('guest');
 Route::get('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
 
@@ -33,14 +33,15 @@ Route::get('/card', [CardController::class,'index'])->middleware('auth');
 Route::get('/item', [ItemController::class,'index'])->middleware('auth');
 Route::get('/trainer', [TrainerController::class,'index'])->middleware('auth');
 Route::get('/home/{id}', [MemberController::class,'getId'])->middleware('auth');
-Route::get('/add-member', [MemberController::class,'create'])->middleware('auth');
-Route::get('/soft-delete-member', [MemberController::class,'softDelete'])->middleware('auth');
-Route::get('/edit-member/{id}', [MemberController::class,'edit'])->middleware('auth');
-Route::get('/member/{id}/restore', [MemberController::class,'restore'])->middleware('auth');
 
-Route::post('/member', [MemberController::class,'store'])->middleware('auth');
-Route::put('/member/{id}', [MemberController::class,'update'])->middleware('auth');
-Route::delete('/member/{id}', [MemberController::class,'destroy'])->middleware('auth');
+Route::get('/add-member', [MemberController::class,'create'])->middleware(['auth', 'admin']);
+Route::get('/soft-delete-member', [MemberController::class,'softDelete'])->middleware(['auth', 'admin']);
+Route::get('/edit-member/{id}', [MemberController::class,'edit'])->middleware(['auth', 'admin']);
+Route::get('/member/{id}/restore', [MemberController::class,'restore'])->middleware(['auth', 'admin']);
+
+Route::post('/member', [MemberController::class,'store'])->middleware(['auth', 'admin']);
+Route::put('/member/{id}', [MemberController::class,'update'])->middleware(['auth', 'admin']);
+Route::delete('/member/{id}', [MemberController::class,'destroy'])->middleware(['auth', 'admin']);
